add it New evenet data, add it delete data and the cron

+++ lib/app.ts
@@ -37,7 +37,7 @@ class App {
             });
     }
     private scheduleCronJob(): void {
-        cron.schedule('29 13  * * *', () => {
+        cron.schedule('29 21  * * *', () => {
             const lockFilePath = './EventDates.lock';
 
             // Check if lock file exists
@@ -69,7 +69,7 @@ class App {
                 fs.unlinkSync(lockFilePath);
             });
         });
-        cron.schedule('00 14  * * *', () => {
+        cron.schedule('00 22  * * *', () => {
             const lockFilePath = './EventLinks.lock';
 
             // Check if lock file exists
@@ -102,7 +102,7 @@ class App {
             });
         });
 
-        cron.schedule('30 14 * * * *', () => {
+        cron.schedule('30 22 * * *', () => {
             const lockFilePath = './EventData.lock';
 
             // Check if lock file exists
@@ -130,6 +130,39 @@ class App {
             // Handle script exit
             scraperProcess.on('close', (code) => {
                 console.log(`EventData.js script exited with code ${code}`);
+                // Remove lock file
+                fs.unlinkSync(lockFilePath);
+            });
+        });
+
+        cron.schedule('0 21 * * *', () => {
+            const lockFilePath = './deleteData.lock';
+
+            // Check if lock file exists
+            if (fs.existsSync(lockFilePath)) {
+                console.log('Scraper already running. Skipping execution.');
+                return;
+            }
+
+            // Create lock file
+            fs.writeFileSync(lockFilePath, '');
+
+            console.log('Running deleteData.js script...');
+            // Execute the results_basketball.js script using child_process.spawn
+            const scraperProcess = childProcess.spawn('node', ['./lib/cron/deleteData.js']);
+
+            // Log stdout and stderr from the results_basketball.js script
+            scraperProcess.stdout.on('data', (data) => {
+                console.log(`stdout: ${data}`);
+            });
+
+            scraperProcess.stderr.on('data', (data) => {
+                console.error(`stderr: ${data}`);
+            });
+
+            // Handle script exit
+            scraperProcess.on('close', (code) => {
+                console.log(`deleteData.js script exited with code ${code}`);
                 // Remove lock file
                 fs.unlinkSync(lockFilePath);
             });
